refactor: move graphql and typeorm configs back into app module file
